# ext/bcmath: Using Karatsuba algorithm in bcmul