… ext/bcmath/tests/bcmul_very_large_numbers.phpt