take2 of https://github.com/php/php-src/pull/14376

This is an implementation of Karatsuba algorithm for mul. It's probably easier to see if check them in commit order. However, notes that there are also changes that will eventually be removed.

One commit would be quite complicated, so I divided it into four major stages of commit.

1. Minimal code implementation that ignores array size and memory allocation efficiency
2. Optimized array size from state 1
3. Optimized memory allocation state from state 2
4. Set the threshold considering the case where normal multiplication is faster (calculation of small values)

# benchmark

I used the following PR code for the benchmark.
https://github.com/php/php-src/pull/14376

When I tried using `UNEXPECTED`, the branching was optimized, and the overall speed became faster.

before:
```
# hyperfine "php /bc/mul/1.php" --warmup 10
Benchmark 1: php /bc/mul/1.php
  Time (mean ± σ):     648.7 ms ±   8.0 ms    [User: 644.4 ms, System: 3.4 ms]
  Range (min … max):   638.1 ms … 660.9 ms    10 runs

# hyperfine "php /bc/mul/2.php" --warmup 10
Benchmark 1: php /bc/mul/2.php
  Time (mean ± σ):     535.1 ms ±   7.6 ms    [User: 529.9 ms, System: 4.3 ms]
  Range (min … max):   526.8 ms … 550.8 ms    10 runs

# hyperfine "php /bc/mul/3.php" --warmup 10
Benchmark 1: php /bc/mul/3.php
  Time (mean ± σ):     552.3 ms ±   5.8 ms    [User: 548.0 ms, System: 3.4 ms]
  Range (min … max):   544.3 ms … 564.5 ms    10 runs

# hyperfine "php /bc/mul/4.php" --warmup 10
Benchmark 1: php /bc/mul/4.php
  Time (mean ± σ):     511.1 ms ±  26.8 ms    [User: 505.5 ms, System: 4.4 ms]
  Range (min … max):   483.8 ms … 558.2 ms    10 runs
```

after (`UNEXPECTED` is unused):
```
# hyperfine "php /bc/mul/1.php" --warmup 10
Benchmark 1: php /bc/mul/1.php
  Time (mean ± σ):     642.9 ms ±   7.9 ms    [User: 638.1 ms, System: 4.0 ms]
  Range (min … max):   631.1 ms … 655.2 ms    10 runs
 
# hyperfine "php /bc/mul/2.php" --warmup 10
Benchmark 1: php /bc/mul/2.php
  Time (mean ± σ):     544.2 ms ±   7.4 ms    [User: 540.5 ms, System: 2.8 ms]
  Range (min … max):   535.8 ms … 554.2 ms    10 runs

# hyperfine "php /bc/mul/3.php" --warmup 10
Benchmark 1: php /bc/mul/3.php
  Time (mean ± σ):     337.8 ms ±   3.7 ms    [User: 332.8 ms, System: 4.1 ms]
  Range (min … max):   331.8 ms … 342.9 ms    10 runs

# hyperfine "php /bc/mul/4.php" --warmup 10
Benchmark 1: php /bc/mul/4.php
  Time (mean ± σ):     216.4 ms ±   5.3 ms    [User: 214.1 ms, System: 1.8 ms]
  Range (min … max):   208.8 ms … 225.4 ms    13 runs
```

after (`UNEXPECTED` is used):
```
# hyperfine "php /bc/mul/1.php" --warmup 10
Benchmark 1: php /bc/mul/1.php
  Time (mean ± σ):     596.6 ms ±   9.9 ms    [User: 592.1 ms, System: 3.6 ms]
  Range (min … max):   585.1 ms … 619.2 ms    10 runs
 
# hyperfine "php /bc/mul/2.php" --warmup 10
Benchmark 1: php /bc/mul/2.php
  Time (mean ± σ):     521.1 ms ±   7.3 ms    [User: 517.3 ms, System: 3.0 ms]
  Range (min … max):   512.8 ms … 533.3 ms    10 runs
 
# hyperfine "php /bc/mul/3.php" --warmup 10
Benchmark 1: php /bc/mul/3.php
  Time (mean ± σ):     338.1 ms ±   6.4 ms    [User: 333.9 ms, System: 3.5 ms]
  Range (min … max):   332.5 ms … 354.3 ms    10 runs
 
# hyperfine "php /bc/mul/4.php" --warmup 10
Benchmark 1: php /bc/mul/4.php
  Time (mean ± σ):     218.6 ms ±   8.6 ms    [User: 214.4 ms, System: 3.5 ms]
  Range (min … max):   207.8 ms … 240.0 ms    13 runs
```

after (Removed adjustment process to multiples of 2):
```
# hyperfine "php /bc/mul/3.php" --warmup 10
Benchmark 1: php /bc/mul/3.php
  Time (mean ± σ):     324.7 ms ±   3.9 ms    [User: 320.8 ms, System: 3.1 ms]
  Range (min … max):   320.2 ms … 331.3 ms    10 runs
 
# hyperfine "php /bc/mul/4.php" --warmup 10
Benchmark 1: php /bc/mul/4.php
  Time (mean ± σ):     210.9 ms ±   5.5 ms    [User: 207.1 ms, System: 3.1 ms]
  Range (min … max):   204.7 ms … 226.1 ms    14 runs
```

# 2 types of thresholds

## BC_MUL_KARATSUBA_DO_CALC_SIZE

If the size becomes smaller than a certain value by repeating the dividing and conquering using the Karatsuba algorithm, normal multiplication is performed without further division.
`BC_MUL_KARATSUBA_DO_CALC_SIZE` specifies the threshold array size. As a result of the measurement, the most efficient value was the same for both 32-bit and 64-bit.

code:
```
$num1 = str_repeat('1234567890', 1024);
$num2 = str_repeat('9876543210', 1024);

for ($i = 0; $i < 500; $i++) {
    bcmul($num1, $num2, 0);
}
```

results:
```
// Karatsuba only
Time (mean ± σ):     397.4 ms ±   6.5 ms    [User: 393.5 ms, System: 3.1 ms]
Range (min … max):   386.3 ms … 405.3 ms    10 runs

// 8
Time (mean ± σ):     232.2 ms ±   7.0 ms    [User: 228.7 ms, System: 2.8 ms]
Range (min … max):   225.4 ms … 246.8 ms    13 runs

// 16 (fastest)
Time (mean ± σ):     209.7 ms ±   3.0 ms    [User: 205.7 ms, System: 3.2 ms]
Range (min … max):   205.1 ms … 216.0 ms    14 runs

// 32
Time (mean ± σ):     221.1 ms ±   5.0 ms    [User: 216.4 ms, System: 4.0 ms]
Range (min … max):   213.5 ms … 229.1 ms    13 runs

// 64
Time (mean ± σ):     256.6 ms ±   2.2 ms    [User: 252.2 ms, System: 3.6 ms]
Range (min … max):   253.7 ms … 260.9 ms    11 runs
```

## BC_MUL_USE_KARATSUBA_SIZE

Denotes the threshold number of digits to use the Karatsuba algorithm for calculations.
The measurement results in 64-bit are as follows.

code:
```
$size = 16; // or 32, 64
$loop = 500000; // or 200000, 100000

$num1 = str_repeat('12345678', $size);
$num2 = str_repeat('98765432', $size);

for ($i = 0; $i < $loop; $i++) {
    bcmul($num1, $num2, 0);
}
```

size = 16, loop = 500000:
```
// standard
Time (mean ± σ):     318.2 ms ±   4.7 ms    [User: 313.3 ms, System: 4.2 ms]
Range (min … max):   310.8 ms … 327.4 ms    10 runs

// karatsuba
Time (mean ± σ):     324.3 ms ±   7.8 ms    [User: 320.2 ms, System: 3.4 ms]
Range (min … max):   316.9 ms … 342.6 ms    10 runs
```

size = 32, loop = 200000:
```
// standard
Time (mean ± σ):     294.7 ms ±   4.5 ms    [User: 289.9 ms, System: 4.0 ms]
Range (min … max):   288.7 ms … 303.8 ms    10 runs

// karatsuba
Time (mean ± σ):     291.1 ms ±   9.4 ms    [User: 286.8 ms, System: 3.5 ms]
Range (min … max):   283.1 ms … 312.4 ms    10 runs
```

size = 64, loop = 100000:
```
// standard
Time (mean ± σ):     401.9 ms ±   5.6 ms    [User: 397.2 ms, System: 3.9 ms]
Range (min … max):   397.2 ms … 413.0 ms    10 runs

// karatsuba
Time (mean ± σ):     346.2 ms ±   5.5 ms    [User: 341.9 ms, System: 3.5 ms]
Range (min … max):   339.1 ms … 355.3 ms    10 runs
```

I also measured 32-bit in the same way (results are omitted).