@SakiTakamachi 2024-06-16
After discussing with Gina I have decided not to make this change, at least not for 8.4.

With the first optimization of multiplication, large values ​​can already be calculated more than 10 times faster.

When I compared the benefits of using the Karatsuba algorithm to further speed things up against the risks of accepting the complex implementation of the Karatsuba algorithm, I decided that the disadvantages probably outweighed the benefits.

If this is really needed, we can try again in the next version.

@Girgias @nielsdos 
Thanks for taking a look at this series of changes!